Scraper/requirements.txt
openpyxl
pandas
python-jobspy